% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylometa_process.R
\name{phylometa_process}
\alias{phylometa_process}
\title{Process phylometa output}
\usage{
phylometa_process(x, groups = 1)
}
\arguments{
\item{x}{output from running Phylometa}

\item{groups}{(integer) number of groups}
}
\description{
Process phylometa output
}
\examples{
file <- system.file("examples/output_eg1.txt", package = "phyrmeta")

}

