% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyrmeta-package.R
\docType{package}
\name{phyrmeta-package}
\alias{phyrmeta}
\alias{phyrmeta-package}
\title{Client to Handle Phylometa Output}
\description{
Client to Handle Phylometa Output
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

