% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylometa_process.R
\name{phylometa_process}
\alias{phylometa_process}
\title{Process phylometa output}
\usage{
phylometa_process(x, groups)
}
\arguments{
\item{x}{(character) a path to a file, the output from running Phylometa.
required.}

\item{groups}{(integer) number of groups. required. This is the number of
groups you specified in your phylometa analysis. if you look at your output
file, it's the number of groups listed in any table}
}
\description{
Process phylometa output
}
\examples{
file <- system.file("examples/example_one_group.txt", package = "phyrmeta")
phylometa_process(file, groups = 1)

file <- system.file("examples/example_two_groups.txt", package = "phyrmeta")
phylometa_process(file, groups = 2)

file <- system.file("examples/example_three_groups.txt", package = "phyrmeta")
phylometa_process(file, groups = 3)

file <- system.file("examples/example_four_groups.txt", package = "phyrmeta")
phylometa_process(file, groups = 4)
}

